Use indirect parametrization from ase
