+++ tests/conftest.py
@@ -2,12 +2,34 @@
 """
 Test configuration
 """
-import ase_fleur  # Make sure the calculator class is registered
+from ase_fleur.calculator import FleurProfile, Fleur
+from ase.test.factories import factory
 from ase.test.factories import Factories
 
 import pytest
 
-version = ase_fleur.__version__
+@factory("fleur")
+class FleurFactory:
+    """
+    Factory for use in ase tests of the Calculator class
+    """
+
+    def __init__(self, executable, inpgen_executable):
+        self.executable = executable
+        self.inpgen_executable = inpgen_executable
+
+    def _profile(self):
+        return FleurProfile([self.executable], [self.inpgen_executable])
+
+    def version(self):
+        self._profile().version()
+
+    def calc(self, **kwargs):
+        return Fleur(profile=self._profile(), **kwargs)
+
+    @classmethod
+    def fromconfig(cls, config):
+        return cls(config.executables["fleur"], config.executables["fleur_inpgen"])
 
 
 def pytest_addoption(parser):
@@ -35,6 +57,12 @@ def pytest_collection_modifyitems(session, config, items):
 
         config.hook.pytest_deselected(items=items)
 
+def pytest_generate_tests(metafunc):
+    """
+    Parametrize calculator factories
+    """
+    from ase.test.factories import parametrize_calculator_tests
+    parametrize_calculator_tests(metafunc)
 
 @pytest.fixture(scope="session")
 def generate_factories(pytestconfig):
